+++ src/js/routes/profile/profile.module.js
@@ -4,7 +4,7 @@ import Checked from "./filters/checked.filter";
 import ModwatchLimitTo from "./filters/modwatchLimitTo.filter";
 
 angular.module("modwatch.profile", [])
-  .controller("ProfileController", ProfileController)
-  .filter("capitalize", Capitalize)
-  .filter("checked", Checked)
-  .filter("modwatchLimitTo", ModwatchLimitTo);
+.controller("ProfileController", ProfileController)
+.filter("capitalize", Capitalize)
+.filter("checked", Checked)
+.filter("modwatchLimitTo", ModwatchLimitTo);